refactor(core): rename vss_core.vst to vss_core.vios
Mechanical rename, no behaviour change. The module is the client for the
VIOS service; "vst" was the older service name, and the CLI surface being
built on it is `vss vios`.

Import order shifted in two files because `vios` sorts before `vlm` where
`vst` sorted after; applied `ruff check --fix` to those.

The `"vst"` service key in the deployment config and the `/vst` ingress
mount are deliberately untouched — those name the route, not the module.
